+++ src/sdax/__init__.py
@@ -26,7 +26,7 @@ from .tasks import (
     task_sync_func,
 )
 
-__version__ = "0.7.0"
+__version__ = "0.7.1"
 
 __all__ = [
     "AsyncTask",